+++ src/main/kotlin/com/mohang/meeting/application/usecase/participant/exception/NoAuthorityParticipatingMeeting.kt
@@ -1,4 +1,4 @@
-package com.mohang.meeting.application.participant.exception
+package com.mohang.meeting.application.usecase.participant.exception
 
 /**
  * Created by ShinD on 2022/09/11.
